Refactor: Centralize command handlers into a single file
