+++ discord/interactions/index.ts
@@ -1,213 +1,20 @@
 import { REST } from "@discordjs/rest";
+import { Client } from "discord.js";
+
 import {
-  ButtonInteraction,
-  ChatInputCommandInteraction,
-  Client,
-  ModalSubmitInteraction,
-  StringSelectMenuInteraction,
-} from "discord.js";
+  ButtonHandler,
+  ChatInputHandler,
+  interactionHandlers,
+  ModalSubmitHandler,
+  SelectMenuHandler,
+} from "../commands/index";
 
 import config from "../../config";
 import WatchTowerLogger from "../../watchTower";
 
-import {
-  handleBackButton,
-  handleConfigCancelCommand,
-  handleConfigConfirmCommand,
-  handleInitialConfigCommand,
-  handleModuleConfigCommand,
-  handleModuleTypeConfigCommand,
-  handleNetworkConfigCommand,
-} from "./addConfigCommand";
-import {
-  handleConnectCommand,
-  handleReconnectNetworkCommand,
-  handleUserNetworkConfigCommand,
-} from "./connectCommand";
-import { handleDebugUserCommand } from "./debugUser";
-import {
-  handleDeleteConfigCommand,
-  handleDeleteConfigConfirmCommand,
-} from "./deleteConfigCommand";
-import {
-  handleDisconnectCommand,
-  handleDisconnectConfirmCommand,
-} from "./disconnectCommand";
-import { handleAnalyticsCommand } from "./handleAnalyticsCommand";
-import { handleHelpCommand } from "./helpCommand";
-import { handleListConfigsCommand } from "./listConfigs";
-import { handleRefreshCommand } from "./refreshCommand";
-import {
-  handleSetConfigCustomApiCommand,
-  handleSetConfigCustomApiModalInput,
-  handleSetConfigCustomApiNext,
-  handleSetConfigCustomApiSelected,
-} from "./setConfigCustomApi";
-import { handleViewConfigCommand } from "./viewConfig";
-
 export const restClient = new REST({ version: "10" }).setToken(
   config.DISCORD_BOT_TOKEN
 );
-
-type ChatInputHandler = (
-  interaction: ChatInputCommandInteraction,
-  client: Client,
-  rest: REST
-) => Promise<void>;
-
-type ButtonHandler = (
-  interaction: ButtonInteraction,
-  client: Client,
-  rest: REST
-) => Promise<void>;
-
-type SelectMenuHandler = (
-  interaction: StringSelectMenuInteraction,
-  client: Client,
-  rest: REST
-) => Promise<void>;
-
-type ModalSubmitHandler = (
-  interaction: ModalSubmitInteraction,
-  client: Client,
-  rest: REST
-) => Promise<void>;
-
-interface HandlerConfig {
-  type: "chatInput" | "button" | "selectMenu" | "modalSubmit";
-  identifier: string;
-  handler:
-    | ChatInputHandler
-    | ButtonHandler
-    | SelectMenuHandler
-    | ModalSubmitHandler;
-  matchType?: "exact" | "startsWith";
-}
-
-const interactionHandlers: HandlerConfig[] = [
-  {
-    type: "chatInput",
-    identifier: "starky-add-config",
-    handler: handleInitialConfigCommand,
-  },
-  {
-    type: "chatInput",
-    identifier: "starky-delete-config",
-    handler: handleDeleteConfigCommand,
-  },
-  {
-    type: "chatInput",
-    identifier: "starky-connect",
-    handler: handleConnectCommand,
-  },
-  {
-    type: "chatInput",
-    identifier: "starky-disconnect",
-    handler: handleDisconnectCommand,
-  },
-  {
-    type: "chatInput",
-    identifier: "starky-refresh",
-    handler: handleRefreshCommand,
-  },
-  {
-    type: "chatInput",
-    identifier: "starky-view-config",
-    handler: handleViewConfigCommand,
-  },
-  {
-    type: "chatInput",
-    identifier: "starky-debug-user",
-    handler: handleDebugUserCommand,
-  },
-  {
-    type: "chatInput",
-    identifier: "starky-set-config-custom-api",
-    handler: handleSetConfigCustomApiCommand,
-  },
-  {
-    type: "chatInput",
-    identifier: "help",
-    handler: handleHelpCommand,
-  },
-  {
-    type: "chatInput",
-    identifier: "list-configs",
-    handler: handleListConfigsCommand,
-  },
-  {
-    type: "chatInput",
-    identifier: "starky-analytics",
-    handler: handleAnalyticsCommand,
-  },
-  {
-    type: "selectMenu",
-    identifier: "starky-config-network",
-    handler: handleNetworkConfigCommand,
-  },
-  {
-    type: "selectMenu",
-    identifier: "starky-config-module-type",
-    handler: handleModuleTypeConfigCommand,
-  },
-  {
-    type: "selectMenu",
-    identifier: "user-config-network",
-    handler: handleUserNetworkConfigCommand,
-  },
-  {
-    type: "selectMenu",
-    identifier: "delete-config-confirm",
-    handler: handleDeleteConfigConfirmCommand,
-  },
-  {
-    type: "selectMenu",
-    identifier: "set-config-custom-api-selected",
-    handler: handleSetConfigCustomApiSelected,
-  },
-  {
-    type: "modalSubmit",
-    identifier: "starky-config-module-config",
-    handler: handleModuleConfigCommand,
-  },
-  {
-    type: "modalSubmit",
-    identifier: "set-config-custom-api-modal",
-    handler: handleSetConfigCustomApiModalInput,
-  },
-  {
-    type: "button",
-    identifier: "starcord-config-confirm",
-    handler: handleConfigConfirmCommand,
-  },
-  {
-    type: "button",
-    identifier: "starcord-config-cancel",
-    handler: handleConfigCancelCommand,
-  },
-  {
-    type: "button",
-    identifier: "disconnect-confirm",
-    handler: handleDisconnectConfirmCommand,
-  },
-  {
-    type: "button",
-    identifier: "set-config-custom-api-next",
-    handler: handleSetConfigCustomApiNext,
-  },
-  {
-    type: "button",
-    identifier: "starky-config-back-",
-    handler: handleBackButton,
-    matchType: "startsWith",
-  },
-  {
-    type: "button",
-    identifier: "reconnect_",
-    handler: handleReconnectNetworkCommand,
-    matchType: "startsWith",
-  },
-];
 
 export const setupInteractions = (client: Client) => {
   client.on("interactionCreate", async (interaction) => {
