+++ discord/slashCommands.ts
@@ -3,89 +3,7 @@ import { Routes, SlashCommandBuilder } from "discord.js";
 
 import config from "../config";
 import WatchTowerLogger from "../watchTower";
-
-type SlashCommandOption = {
-  type: "role" | "user";
-  name: string;
-  description: string;
-  required: boolean;
-};
-
-type SlashCommandData = {
-  name: string;
-  description: string;
-  options?: SlashCommandOption[];
-};
-
-export const slashCommandsArray: SlashCommandData[] = [
-  {
-    name: "starky-connect",
-    description: "Connect your Starknet wallet to this Discord server",
-  },
-  {
-    name: "starky-disconnect",
-    description: "Disconnect your Starknet wallet from this Discord server",
-  },
-  {
-    name: "starky-add-config",
-    description: "Add a starky configuration to this server",
-    options: [
-      {
-        type: "role",
-        name: "role",
-        description:
-          "What role do you want to assign to people matching your criteria?",
-        required: true,
-      },
-    ],
-  },
-  {
-    name: "starky-delete-config",
-    description: "Delete a Starky configuration from this server",
-  },
-  {
-    name: "starky-refresh",
-    description: "Refresh your Starky roles on this server",
-  },
-  {
-    name: "starky-view-config",
-    description: "View your Starky configurations on this server",
-    options: [
-      {
-        type: "role",
-        name: "role",
-        description: "The role of the configuration you want to check.",
-        required: true,
-      },
-    ],
-  },
-  {
-    name: "starky-debug-user",
-    description: "Debug a user on this server",
-    options: [
-      {
-        type: "user",
-        name: "user",
-        description: "The user you want to debug.",
-        required: true,
-      },
-    ],
-  },
-  {
-    name: "starky-set-config-custom-api",
-    description:
-      "Set a custom API URI for a configuration (Instead of Starkscan)",
-  },
-  { name: "help", description: "Get help with Starky" },
-  {
-    name: "list-configs",
-    description: "View your Starky configurations on this server",
-  },
-  {
-    name: "starky-analytics",
-    description: "View analytics for this server",
-  },
-];
+import { slashCommandsArray, SlashCommandData } from "./commands/index";
 
 const buildSlashCommand = (
   commandData: SlashCommandData
